requirements.txt
fastapi== 0.110.1
uvicorn==0.29.0
openai==1.17.0
langchain==0.1.14
langchain_community==0.0.32
langchain_core==0.1.41
langchain_openai==0.1.2
langchain_text-splitters==0.0.1
pydantic==2.6.4
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
faiss-cpu==1.8.0
uvicorn==0.29.0
